chore: use generated types from vega protos - move out some type from wallet to deal-ticket module
