+++ libs/deal-ticket/src/components/deal-ticket/deal-ticket-container.tsx
@@ -1,12 +1,44 @@
 import { AsyncRenderer, Splash } from '@vegaprotocol/ui-toolkit';
 import { t } from '@vegaprotocol/i18n';
 import { useThrottledDataProvider } from '@vegaprotocol/data-provider';
+import type { OrderSubmission } from '@vegaprotocol/wallet';
 import {
-  convertDealTicketToOrderSubmission,
   useVegaTransactionStore,
+  SideMap,
+  TimeInForceMap,
+  TypeMap,
 } from '@vegaprotocol/wallet';
 import { useMarket, marketDataProvider } from '@vegaprotocol/markets';
 import { DealTicket } from './deal-ticket';
+import type * as Schema from '@vegaprotocol/types';
+
+export interface DealTicketOrderSubmission {
+  marketId: string;
+  reference?: string;
+  type: Schema.OrderType;
+  side: Schema.Side;
+  timeInForce: Schema.OrderTimeInForce;
+  size: string;
+  price?: string;
+  expiresAt?: string;
+  postOnly?: boolean;
+  reduceOnly?: boolean;
+}
+
+export const convertDealTicketToOrderSubmission = (
+  dealTicketOrder: DealTicketOrderSubmission
+) => {
+  return {
+    ...dealTicketOrder,
+    expiresAt: dealTicketOrder.expiresAt
+      ? BigInt(dealTicketOrder.expiresAt)
+      : null,
+    size: BigInt(dealTicketOrder.size),
+    side: SideMap[dealTicketOrder.side],
+    timeInForce: TimeInForceMap[dealTicketOrder.timeInForce],
+    type: TypeMap[dealTicketOrder.type],
+  } as OrderSubmission;
+};
 
 export interface DealTicketContainerProps {
   marketId: string;
