+++ libs/wallet/src/use-vega-transaction-store.spec.tsx
@@ -13,9 +13,8 @@ import type {
   OrderCancellation,
   OrderCancellationBody,
   WithdrawSubmissionBody,
-} from './connectors/vega-connector';
-import type { DealTicketOrderAmendment } from './utils';
-import { convertDealTicketToOrderAmendment } from './utils';
+  OrderAmendment,
+} from './connectors';
 
 jest.mock('./utils', () => ({
   ...jest.requireActual('./utils'),
@@ -38,14 +37,14 @@ describe('useVegaTransactionStore', () => {
     } as unknown as WithdrawSubmission,
   };
 
-  const orderAmendment = {
+  const orderAmendment: OrderAmendment = {
     orderId: '6a4fcd0ba478df2f284ef5f6d3c64a478cb8043d3afe36f66f92c0ed92631e64',
     marketId:
       '3aa2a828687cc3d59e92445d294891cbbd40e2165bbfb15674158ef5d4e8848d',
     price: '1122',
-    timeInForce: OrderTimeInForce.TIME_IN_FORCE_GTC,
-    sizeDelta: 0,
-  } as DealTicketOrderAmendment;
+    timeInForce: vegaProtos.Order.TimeInForce.TIME_IN_FORCE_GTC,
+    sizeDelta: BigInt(0),
+  };
 
   const originalOrder = {
     type: OrderType.TYPE_LIMIT,
@@ -104,22 +103,13 @@ describe('useVegaTransactionStore', () => {
   it('updates an order with order amendment', () => {
     useVegaTransactionStore
       .getState()
-      .create(
-        { orderAmendment: convertDealTicketToOrderAmendment(orderAmendment) },
-        originalOrder
-      );
+      .create({ orderAmendment }, originalOrder);
     useVegaTransactionStore
       .getState()
-      .create(
-        { orderAmendment: convertDealTicketToOrderAmendment(orderAmendment) },
-        originalOrder
-      );
+      .create({ orderAmendment }, originalOrder);
     useVegaTransactionStore
       .getState()
-      .create(
-        { orderAmendment: convertDealTicketToOrderAmendment(orderAmendment) },
-        originalOrder
-      );
+      .create({ orderAmendment }, originalOrder);
     const transaction = useVegaTransactionStore.getState().transactions[1];
     useVegaTransactionStore
       .getState()
@@ -147,7 +137,6 @@ describe('useVegaTransactionStore', () => {
         price: '1122',
         sizeDelta: BigInt(0),
         timeInForce: vegaProtos.Order.TimeInForce.TIME_IN_FORCE_GTC,
-        expiresAt: null,
       },
     };
     expect(
